+++ src/GRA.Domain.Repository/IUserRepository.cs
@@ -16,6 +16,7 @@ namespace GRA.Domain.Repository
         Task<User> GetByUsernameAsync(string username);
         Task<int> GetCountAsync(UserFilter filter);
         Task<int> GetCountAsync(ReportCriterion request);
+        Task<int> GetFirstTimeCountAsync(ReportCriterion request);
         Task<int> GetHouseholdCountAsync(int householdHeadUserId);
         Task<DataWithId<IEnumerable<string>>> GetUserIdAndUsernames(string email);
         Task<IEnumerable<Model.User>> PageAllAsync(UserFilter filter);